+++ src/main/java/com/jb/CouponSystem3/controllers/CompanyControllers.java
@@ -5,6 +5,7 @@ import com.jb.CouponSystem3.beans.Company;
 import com.jb.CouponSystem3.beans.Coupon;
 import com.jb.CouponSystem3.exceptions.CouponSecurityException;
 import com.jb.CouponSystem3.exceptions.CouponSystemException;
+import com.jb.CouponSystem3.exceptions.SecMsg;
 import com.jb.CouponSystem3.security.*;
 import com.jb.CouponSystem3.services.CompanyService;
 import lombok.RequiredArgsConstructor;
@@ -34,13 +35,16 @@ public class CompanyControllers extends ClientController {
         String password = loginRequest.getPassword();
         ClientType type = loginRequest.getType();
         UUID token = loginManager.loginUUID(email, password, type);
-        return new LoginResponse(token);
+        return new LoginResponse(token, email);
     }
 
     @PostMapping("coupons")
     @ResponseStatus(HttpStatus.CREATED)
     void addCoupon(@RequestHeader("Authorization") UUID token, @RequestBody Coupon coupon) throws CouponSystemException, CouponSecurityException {
         int companyId = tokenManager.getUserId(token);
+        if (tokenManager.getType(token) != ClientType.COMPANY) {
+            throw new CouponSecurityException(SecMsg.INVALID_TOKEN);
+        }
         companyService.addCoupon(companyId, coupon);
     }
 
@@ -48,6 +52,9 @@ public class CompanyControllers extends ClientController {
     @ResponseStatus(HttpStatus.NO_CONTENT)
     void updateCoupon(@RequestHeader("Authorization") UUID token, @PathVariable int couponId, @RequestBody Coupon coupon) throws CouponSystemException, CouponSecurityException {
         int companyId = tokenManager.getUserId(token);
+        if (tokenManager.getType(token) != ClientType.COMPANY) {
+            throw new CouponSecurityException(SecMsg.INVALID_TOKEN);
+        }
         companyService.updateCoupon(companyId, couponId, coupon);
     }
 
@@ -55,48 +62,72 @@ public class CompanyControllers extends ClientController {
     @ResponseStatus(HttpStatus.NO_CONTENT)
     void deleteCoupon(@RequestHeader("Authorization") UUID token, @PathVariable int couponId) throws CouponSystemException, CouponSecurityException {
         int companyId = tokenManager.getUserId(token);
+        if (tokenManager.getType(token) != ClientType.COMPANY) {
+            throw new CouponSecurityException(SecMsg.INVALID_TOKEN);
+        }
         companyService.deleteCoupon(companyId, couponId);
     }
 
     @GetMapping("coupons")
     List<Coupon> getAllCouponsByCompanyId(@RequestHeader("Authorization") UUID token) throws CouponSecurityException {
         int companyId = tokenManager.getUserId(token);
+        if (tokenManager.getType(token) != ClientType.COMPANY) {
+            throw new CouponSecurityException(SecMsg.INVALID_TOKEN);
+        }
         return companyService.getAllCouponsByCompanyId(companyId);
     }
 
     @GetMapping("coupons/{couponId}")
     Coupon getOneCouponById(@RequestHeader("Authorization") UUID token, @PathVariable int couponId) throws CouponSystemException, CouponSecurityException {
         int companyId = tokenManager.getUserId(token);
+        if (tokenManager.getType(token) != ClientType.COMPANY) {
+            throw new CouponSecurityException(SecMsg.INVALID_TOKEN);
+        }
         return companyService.getOneCouponByIdAndCouponId(companyId, couponId);
     }
 
     @GetMapping("coupons2/{couponId}")
     Coupon getOneCouponByIdAndCouponId(@RequestHeader("Authorization") UUID token, @PathVariable int couponId) throws CouponSystemException, CouponSecurityException {
         int companyId = tokenManager.getUserId(token);
+        if (tokenManager.getType(token) != ClientType.COMPANY) {
+            throw new CouponSecurityException(SecMsg.INVALID_TOKEN);
+        }
         return companyService.getOneCouponByIdAndCouponId(companyId, couponId);
     }
 
     @GetMapping("coupons2")
     List<Coupon> getCompanyCoupons(@RequestHeader("Authorization") UUID token) throws CouponSecurityException {
         int companyId = tokenManager.getUserId(token);
+        if (tokenManager.getType(token) != ClientType.COMPANY) {
+            throw new CouponSecurityException(SecMsg.INVALID_TOKEN);
+        }
         return companyService.getCompanyCoupons(companyId);
     }
 
     @GetMapping("coupons/category")
     List<Coupon> getCompanyCouponsByCategory(@RequestHeader("Authorization") UUID token, @RequestParam Category category) throws CouponSecurityException {
         int companyId = tokenManager.getUserId(token);
+        if (tokenManager.getType(token) != ClientType.COMPANY) {
+            throw new CouponSecurityException(SecMsg.INVALID_TOKEN);
+        }
         return companyService.getCompanyCoupons(companyId, category);
     }
 
-    @GetMapping("coupons/maxPrice")
-    List<Coupon> getCompanyCouponsByMaxPrice(@RequestHeader("Authorization") UUID token, @RequestParam double maxPrice) throws CouponSecurityException {
+    @GetMapping("coupons/price/max")
+    List<Coupon> getCompanyCouponsByMaxPrice(@RequestHeader("Authorization") UUID token, @RequestParam double value) throws CouponSecurityException {
         int companyId = tokenManager.getUserId(token);
-        return companyService.getCompanyCoupons(companyId, maxPrice);
+        if (tokenManager.getType(token) != ClientType.COMPANY) {
+            throw new CouponSecurityException(SecMsg.INVALID_TOKEN);
+        }
+        return companyService.getCompanyCoupons(companyId, value);
     }
 
     @GetMapping("details")
     Company getCompanyDetails(@RequestHeader("Authorization") UUID token) throws CouponSystemException, CouponSecurityException {
         int companyId = tokenManager.getUserId(token);
+        if (tokenManager.getType(token) != ClientType.COMPANY) {
+            throw new CouponSecurityException(SecMsg.INVALID_TOKEN);
+        }
         return companyService.getCompanyDetails(companyId);
     }
